Bug 2000658 - Strip "rcm" parameter on LinkedIn with "Copy Clean Link"
Add "rcm" to the list of query parameters stripped from LinkedIn URLs
when using the "Copy Clean Link" feature. Also add www.linkedin.com to
the list of LinkedIn domains to ensure coverage of the main domain.

The "rcm" parameter is used by LinkedIn for tracking purposes and should
be removed when sharing links.